+++ pkg/tfpfbridge/info.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tfbridge
+package tfpfbridge
 
 import (
 	"github.com/hashicorp/terraform-plugin-framework/provider"
