+++ pkg/tfpfbridge/provider_diffconfig.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tfbridge
+package tfpfbridge
 
 import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
